pyproject.toml
[tool.poetry]
name = "LNWzZ"
version = "0.1.0"
description = ""
authors = ["lst97 <laisiotou1997@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
discord = "^1.7.3"
aiohttp-session = "^2.9.0"
python-dotenv = "^0.19.0"
cryptography = "^3.4.8"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.8b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
